experiments/partitioned-heat-conduction/solver-fenics/requirements.txt
numpy >1, <2
fenicsprecice~=2.0
scipy

# Assuming FEniCS from ppa:fenics-packages/fenics was installed https://fenicsproject.org/download/archive/
# Use --system-site-packages in venv
fenics-dijitso~=2019.0
fenics-dolfin~=2019.0
fenics-ffc~=2019.0
fenics-fiat~=2019.0
fenics-ufl-legacy~=2022.0

# For convergence studies
pandas
jinja2
GitPython
prepesthel==0.1.1  # custom packages developed by Benjamin Rodenberg
